# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_controllers.errorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Back end for 50 Cent Application
  title: 50 Cent API
  version: "1.1"
paths:
  /admin/actions/users:
    get:
      consumes:
      - application/json
      description: Get all unverified users (both investors and loaners) in JSON
      operationId: admin-get-all-unverified-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all unverified users in JSON
      tags:
      - Admin
  /admin/actions/users/{accountType}/{id}/decline:
    post:
      description: Decline account of user with given ID
      operationId: admin-decline-user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Account type(either 'consumers' or 'investors')
        in: path
        name: accountType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Decline user with given ID
      tags:
      - Admin
  /admin/actions/users/consumers/{id}/approve:
    post:
      description: Approve account of user with given ID
      operationId: admin-approve-consumer
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Approve consumer with given ID
      tags:
      - Admin
  /admin/actions/users/investors/{id}/approve:
    post:
      description: Approve account of user with given ID
      operationId: admin-approve-investor
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Approve investor with given ID
      tags:
      - Admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"